pycdf.istp: move non-ISTP-specific into pycdf; docstring formatting
