+++ spacepy/pycdf/istp.py
@@ -24,7 +24,6 @@ This module supports that subset of CDFs.
 
     fillval
     format
-    get_minmax
 """
 
 import datetime
@@ -33,8 +32,9 @@ import os.path
 import re
 
 import numpy
-import spacepy.pycdf.const
 import spacepy.datamodel
+import spacepy.pycdf
+import spacepy.pycdf.const
 
 
 class VariableChecks(object):
@@ -181,7 +181,7 @@ class VariableChecks(object):
         errs = []
         raw_v = v.cdf_file.raw_var(v.name())
         data = raw_v[...]
-        minval, maxval = get_minmax(raw_v.type())
+        minval, maxval = spacepy.pycdf.lib.get_minmax(raw_v.type())
         if 'VALIDMIN' in raw_v.attrs:
             idx = data < raw_v.attrs['VALIDMIN']
             if ('FILLVAL' in raw_v.attrs) and (idx.size != 0):
@@ -229,7 +229,7 @@ class VariableChecks(object):
         """
         errs = []
         raw_v = v.cdf_file.raw_var(v.name())
-        minval, maxval = get_minmax(raw_v.type())
+        minval, maxval = spacepy.pycdf.lib.get_minmax(raw_v.type())
         data = raw_v[...]
         if 'SCALEMIN' in raw_v.attrs:
             if (raw_v.attrs['SCALEMIN'] < minval) or \
@@ -568,42 +568,3 @@ def format(v, use_scaleminmax=False, dryrun=False):
         v.attrs.new('FORMAT', data=fmt, type=spacepy.pycdf.const.CDF_CHAR)
 
 
-def get_minmax(cdftype):
-    """Find minimum, maximum possible value based on CDF type.
-
-    This returns the "raw" (unparsed) value.
-
-    :param int cdftype: CDF type number from :mod:`~spacepy.pycdf.const`
-    :return: minimum, maximum value supported by type
-    :rtype: tuple
-    """
-    try:
-        cdftype = cdftype.value
-    except:
-        pass
-    if cdftype == spacepy.pycdf.const.CDF_EPOCH.value:
-        return (
-            spacepy.pycdf.lib.datetime_to_epoch(
-                datetime.datetime(1, 1, 1)),
-            #Can get asymptotically closer, but why bother
-            spacepy.pycdf.lib.datetime_to_epoch(
-                datetime.datetime(9999, 12, 31, 23, 59, 59)))
-    elif cdftype == spacepy.pycdf.const.CDF_EPOCH16.value:
-        return (
-            spacepy.pycdf.lib.datetime_to_epoch16(
-                datetime.datetime(1, 1, 1)),
-            spacepy.pycdf.lib.datetime_to_epoch16(
-                datetime.datetime(9999, 12, 31, 23, 59, 59)))
-    elif cdftype == spacepy.pycdf.const.CDF_TIME_TT2000.value:
-        inf = numpy.iinfo(numpy.int64)
-        return (inf.min, inf.max)
-    dtype = spacepy.pycdf.lib.numpytypedict.get(cdftype, None)
-    if dtype is None:
-        raise ValueError('Unknown data type: {}'.format(cdftype))
-    if numpy.issubdtype(dtype, numpy.integer):
-        inf = numpy.iinfo(dtype)
-    elif numpy.issubdtype(dtype, numpy.float):
-        inf = numpy.finfo(dtype)
-    else:
-        raise ValueError('Unknown data type: {}'.format(cdftype))
-    return (inf.min, inf.max)
